requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
Pillow==10.2.0
python-dotenv==1.0.1
faster-whisper==1.0.3
ctranslate2==4.3.1
sentence-transformers==2.6.1
chromadb==0.4.24
anthropic==0.20.0
openai==1.14.0
torch==2.2.1
torchaudio==2.2.1
channels==4.0.0
channels-redis==4.2.0
celery==5.3.6
redis==5.0.3
django-filter==23.5
numpy==1.26.4
scikit-learn==1.4.2
pydub==0.25.1
ffmpeg-python==0.2.0